Extract separate modules for S3TestCase and MirrorTestCase
to avoid circular imports between test/indexer/__init__.py and test/service/__init__.py
