+++ test/service/__init__.py
@@ -13,12 +13,9 @@ from typing import (
     ClassVar,
     Optional,
     Union,
-    cast,
-    get_args,
 )
 from unittest.mock import (
     MagicMock,
-    PropertyMock,
     patch,
 )
 import uuid
@@ -30,25 +27,12 @@ from more_itertools import (
     flatten,
     one,
 )
-from moto import (
-    mock_aws,
-)
-from mypy_boto3_s3.client import (
-    S3Client,
-)
-from mypy_boto3_s3.literals import (
-    BucketLocationConstraintType,
-)
 
 from app_test_case import (
     LocalAppTestCase,
 )
 from azul import (
     cached_property,
-    config,
-)
-from azul.deployment import (
-    aws,
 )
 from azul.indexer import (
     Bundle,
@@ -78,11 +62,11 @@ from azul.types import (
     JSON,
     JSONs,
 )
-from azul_test_case import (
-    AzulUnitTestCase,
-)
 from indexer import (
     IndexerTestCase,
+)
+from s3_test_case import (
+    S3TestCase,
 )
 
 log = get_test_logger(__name__)
@@ -220,23 +204,6 @@ class DocumentCloningTestCase(WebServiceTestCase, metaclass=ABCMeta):
                                     doc_type=DocumentType.aggregate))
 
 
-class S3TestCase(AzulUnitTestCase):
-
-    @property
-    def _s3(self) -> S3Client:
-        return aws.s3
-
-    def setUp(self) -> None:
-        super().setUp()
-        self.addPatch(mock_aws())
-
-    def _create_test_bucket(self, bucket_name: str):
-        assert config.region in get_args(BucketLocationConstraintType)
-        location = cast(BucketLocationConstraintType, config.region)
-        self._s3.create_bucket(Bucket=bucket_name,
-                               CreateBucketConfiguration={'LocationConstraint': location})
-
-
 class StorageServiceTestCase(S3TestCase):
     """
     A mixin for test cases that utilize StorageService.
@@ -249,20 +216,6 @@ class StorageServiceTestCase(S3TestCase):
     def setUp(self) -> None:
         super().setUp()
         self._create_test_bucket(self.storage_service.bucket_name)
-
-
-class MirrorTestCase(S3TestCase):
-    mirror_bucket = 'test-mirror-bucket'
-
-    def setUp(self):
-        super().setUp()
-        self.addPatch(patch.object(type(config),
-                                   'enable_mirroring',
-                                   new=PropertyMock(return_value=True)))
-        self.addPatch(patch.object(type(config),
-                                   'mirror_bucket',
-                                   new=PropertyMock(return_value=self.mirror_bucket)))
-        self._create_test_bucket(self.mirror_bucket)
 
 
 @deprecated('Instead of decorating your test case, or its test methods in it, '
